requirements.txt
#Pose Estimation
TensorFlow-gpu==2.4
opencv-python==4.2.0
scipy==1.4.1
slidingwindow
pycocotools
tqdm
sklearn

#Chatbot
speechRecognition
PyAudio

#Face Expression Recognition
dlib
face-recognition

#Robot
numpy==1.19.5
opencv-python==4.2.0
PyQt5